move non-plugin config to core directory
