+++ xdg_config/nvim/lua/ian/core/options.lua
